requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
Django==3.1.2
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-cryptography==1.0
django-taggit==1.3.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
importlib-metadata==2.0.0
Markdown==3.3
markdown-checklist==0.4.3
Pillow==7.2.0
pycparser==2.20
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
reportlab==3.5.53
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.1
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zipp==3.3.0
